requirements.txt
langchain
langgraph
langchain_core
langgraph-checkpoint-sqlite
chromadb
langchain-community
langchain-google-genai
python-dotenv
langchain-ollama
langchain-chroma